+++ Source Code Inspection/my-app/src/main/java/com/mycompany/app/App.java
@@ -1,0 +1,10 @@
+package com.mycompany.app;
+
+
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
