+++ Source Code Inspection/my-app/src/main/java/com/mycompany/app/PapelMoeda.java
@@ -1,0 +1,24 @@
+package com.mycompany.app;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class PapelMoeda {
+
+    private int valor;
+    private int quantidade;
+
+    public PapelMoeda(int valor, int quantidade) {
+        this.valor = valor;
+        this.quantidade = quantidade;
+    }
+
+    public int getValor() {
+        return valor;
+    }
+
+    public int getQuantidade() {
+        return quantidade;
+    }
+}
