+++ Source Code Inspection/my-app/src/main/java/com/mycompany/app/PapelMoedaInvalidaException.java
@@ -1,0 +1,9 @@
+package br.calebe.ticketmachine.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class PapelMoedaInvalidaException extends Exception {
+    
+}
