+++ Source Code Inspection/my-app/src/main/java/com/mycompany/app/SaldoInsuficienteException.java
@@ -1,0 +1,11 @@
+package br.calebe.ticketmachine.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class SaldoInsuficienteException extends Exception {
+    public SaldoInsuficienteException() {
+        super("Saldo insuficiente para completar a operação.");
+    }
+}
