+++ Source Code Inspection/my-app/src/test/java/com/mycompany/app/AppTest.java
@@ -1,0 +1,31 @@
+package com.mycompany.app;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    
+    
+
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        //assertTrue( true );
+    }
+
+    @Test
+    public void getvalorCerto(){
+
+        PapelMoeda papel = new PapelMoeda(10, 5);
+
+        int quantidade = papel.getQuantidade();
+
+        Assert.assertEquals(10, quantidade);
+
+        //assertTrue(true);
+    }
+}
