--- Source Code Inspection/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=35e2c666
-build.xml.script.CRC32=03bb7497
-build.xml.stylesheet.CRC32=28e38971@1.53.1.46
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=5804ec04
-nbproject/build-impl.xml.script.CRC32=b9819ea0
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
