--- Source Code Inspection/src/br/calebe/ticketmachine/exception/PapelMoedaInvalidaException.java
@@ -1,9 +1,0 @@
-package br.calebe.ticketmachine.exception;
-
-/**
- *
- * @author Calebe de Paula Bianchini
- */
-public class PapelMoedaInvalidaException extends Exception {
-    
-}
