--- Source Code Inspection/src/br/calebe/ticketmachine/exception/SaldoInsuficienteException.java
@@ -1,11 +1,0 @@
-package br.calebe.ticketmachine.exception;
-
-/**
- *
- * @author Calebe de Paula Bianchini
- */
-public class SaldoInsuficienteException extends Exception {
-    public SaldoInsuficienteException() {
-        super("Saldo insuficiente para completar a operação.");
-    }
-}
